requirements.txt
coverage==5.5
deal==4.5.0
flake8==3.9.2
pytz==2021.1
responses==0.13.1
requests==2.25.1
pytest==6.2.2
pytest-cov==2.11.1
pytest-mock==3.6.1
pytest-factoryboy==2.1.0
django==3.2.5
factory_boy==3.2.0
Faker==8.10.1
lazy-object-proxy==1.6.0
django-extensions==3.1.3
pytest-django==4.4.0
djangorestframework==3.12.4
mypy==0.910
types-requests==2.25.0
